Move RecentProjects one level up
